requirements.txt
requests~=2.28.2
berserk~=0.10.0
python-dotenv~=0.21.1
Jinja2~=3.1.2